Fix build error on MingW
Gnulib does not currently implement the `F_GETFL` and `F_SETFL` flags on
mingw. As a result building Wget failed on Windows. We don't currently
have a good solution to this problem and Windows users of Wget will not
get the new functionality of non-blocking file I/O reads. The current
solution is simply to provide a Windows specific no-op stub

 * src/mswindows.c (set_fd_nonblocking): Provide an empty stub function
   for Windows systems
 * src/mswindows.h (set_fd_nonblocking): Export the function so it is
   available in util.c where it is needed
 * src/util.c (set_fd_nonblocking): New function to set a fd as
   O_NONBLOCK on non-Windows, non-MSDOS systems
   (wget_read_from_file): Split the O_NONBLOCK code into
   set_fd_nonblocking and make all file I/O non-blocking, not just stdin