Fix a static analysis false positive
Clang's static analyzer believes that it is possible for a realloc
request to return NULL. However, the `xrealloc` implementation used
in Wget will trigger a program crash if enough memory couldn't be
found. I'm not sure why the analyzer isn't able to see that. But let's
pacify it by asserting here that xrealloc will always return a non-null

 * src/wget.h (DO_REALLOC): Assert that xrealloc will always return
   non-null